fix: resolve mangaball parse error for chapter volumes
Update `volume` from Int to Float in `Chapter` class (`Dto.kt`) to support decimal volumes, preventing JSON parse crash.
Update `MangaBall.kt` to strip `.0` suffix for display strings to preserve user experience for whole number volumes.

+++ app/src/main/java/eu/kanade/tachiyomi/source/online/merged/mangaball/Dto.kt
@@ -25,7 +25,7 @@ class Chapter(
     val language: String,
     val group: Group,
     val date: String,
-    val volume: Int,
+    val volume: Float,
 )
 
 @Serializable class Group(@SerialName("_id") val id: String, val name: String)
